+++ src/lib/UI/domainComponents/lists/SortRecipesSelect/SortRecipesSelect.svelte
@@ -1,5 +1,6 @@
 <script context="module" lang="ts">
   import type { Recipe } from "$lib/domain/entities/Recipe";
+  import { RecipesSortOrderEnum } from "$lib/domain/sort/RecipesSortOrderEnum";
   import {
     sortRecipesByTimestampDesc,
     sortRecipesByTimestampAsc,
@@ -8,7 +9,7 @@
     sortRecipesByRatingDesc,
     sortRecipesByRatingAsc,
     sortRecipesByFavorite
-  } from "$lib/domain/helpers/sortRecipes";
+  } from "$lib/domain/sort/sortRecipes";
   import { loadSortOrder, saveSortOrder } from "$lib/localStorage/persistSortOrder";
 
   const SELECT_ID = "recipes-sort-order";
@@ -29,8 +30,6 @@
 </script>
 
 <script lang="ts">
-  import { RecipesSortOrderEnum } from "./RecipesSortOrderEnum";
-
   // Props:
   export let sortOrderValue: {
     value: RecipesSortOrderEnum;
@@ -70,10 +69,8 @@
     justify-content: flex-start;
     align-items: center;
     column-gap: 0.5rem;
-
     /* Typography: */
     @apply text-base font-normal tracking-tight;
-
     /* Colors: */
     @apply text-gray-900 dark:text-white;
   }
@@ -83,7 +80,6 @@
     @apply focus:ring-blue-500 focus:border-blue-500;
     @apply dark:bg-gray-700 dark:border-gray-600 dark:placeholder-gray-400 dark:text-white;
     @apply dark:focus:ring-blue-500 dark:focus:border-blue-500;
-
     padding: 0.625rem 1rem 0.625rem 0.5rem;
   }
 </style>
