+++ src/lib/domain/enhancedEntities/EnhancedCoffeeBeans.ts
