+++ src/routes/coffee_beans/[name]/+page.svelte
@@ -11,13 +11,13 @@
   import { routes } from "$lib/domain/constants/routes";
   import { CoffeeBeans } from "$lib/domain/entities/CoffeeBeans";
   import type { Recipe } from "$lib/domain/entities/Recipe";
-  import { sortRecipesByTimestampDesc } from "$lib/domain/helpers/sortRecipes";
-  import type { Count } from "$lib/types/Count";
+  import { RecipesSortOrderEnum } from "$lib/domain/sort/RecipesSortOrderEnum";
+  import { sortRecipesByTimestampDesc } from "$lib/domain/sort/sortRecipes";
+  import type { Count } from "$lib/helperTypes/Count";
   import RecipeCard from "$lib/UI/domainComponents/cards/RecipeCard.svelte";
   import MyFab from "$lib/UI/domainComponents/FABs/AddRecipeFab.svelte";
   import Loading from "$lib/UI/domainComponents/lists/Loading.svelte";
   import NoItemsYetP from "$lib/UI/domainComponents/lists/NoItemsYetP.svelte";
-  import { RecipesSortOrderEnum } from "$lib/UI/domainComponents/lists/SortRecipesSelect/RecipesSortOrderEnum";
   import SortRecipesSelect from "$lib/UI/domainComponents/lists/SortRecipesSelect/SortRecipesSelect.svelte";
   import DropdownMenu from "$lib/UI/genericComponents/dropdownMenu/DropdownMenu.svelte";
   import DropdownMenuItem from "$lib/UI/genericComponents/dropdownMenu/DropdownMenuItem.svelte";
